refactor(ui): split config form into per-domain field renderers
